separate prepared configurable properties
