+++ Sources/SwiftLayoutUtil/Describe/ConfigurablePropertys/ConfigurableProperties+UIStackView.swift
@@ -1,0 +1,20 @@
+import Foundation
+import UIKit
+
+extension ConfigurableProperties {
+    
+    func registUIStackView() {
+        regist(UIStackView.self, propertiesHandler: uiStackViewDefaultConfigurablePropertys)
+    }
+    
+    private func uiStackViewDefaultConfigurablePropertys(defaultReferenceView stackView: UIStackView) -> [ConfigurableProperty] {
+        return [
+            .property(keypath: \.alignment, defaultReferenceView: stackView) { "$0.alignment = \($0.configuration)"},
+            .property(keypath: \.axis, defaultReferenceView: stackView) { "$0.axis = \($0.configuration)"},
+            .property(keypath: \.distribution, defaultReferenceView: stackView) { "$0.distribution = \($0.configuration)"},
+            .property(keypath: \.isBaselineRelativeArrangement, defaultReferenceView: stackView) { "$0.isBaselineRelativeArrangement = \($0)"},
+            .property(keypath: \.spacing, defaultReferenceView: stackView) { "$0.spacing = \($0)"},
+        ]
+    }
+    
+}
